Trocados tipos de dados de latitude e longitude

+++ shareblood/location/models.py
@@ -14,5 +14,5 @@ class Location(models.Model):
     zipcode = models.CharField(null=True, blank=True, max_length=255)
     street = models.CharField(null=True, blank=True, max_length=255)
     neighborhood = models.CharField(null=True, blank=True, max_length=255)
-    latitude = models.BigIntegerField()
-    longitude = models.BigIntegerField()
+    latitude = models.DecimalField(max_digits=20, decimal_places=8)
+    longitude = models.DecimalField(max_digits=20, decimal_places=8)
